Add ROAMController class which will facilitate testing
Introduce a function (x, y) -> target lod, which will
be used to determine if a simplex needs to be split.
Add code to recursively force split when a simplex is
single_split and a resulting simplex needs to be split
further.

I am not sure that the recursive force splitting works
correctly. It has not been tested.